# message auto-generated for no-merge-commit merge:
!265 merge develop into develop

merge 730 into develop

Created-by: shenwei21
Commit-by: shenwei21
Merged-by: yanghui79
Description: Paired: [GitHub #187](https://github.com/openJiuwen-ai/agent-core-java/pull/187) ↔ [GitCode !265](https://gitcode.com/openJiuwen/agent-core-java/merge_requests/265)

<!--  Thanks for sending a pull request!  Here are some tips for you:

1) If this is your first time, please read our contributor guidelines: https://gitcode.com/openJiuwen/community/blob/master/CONTRIBUTING.md

2) If you want to contribute your code but don't know who will review and merge, please add label `openjiuwen-assistant` to the pull request, we will find and do it as soon as possible.
-->

**What type of PR is this?**
<!-- 
选择下面一种标签替换下方 `/kind <label>`，可选标签类型有：
- /kind bug 
- /kind task
- /kind feature
- /kind refactor
- /kind clean_code
如PR描述不符合规范，修改PR描述后需要/check-pr重新检查PR规范。
-->
/kind <label>

**Self-checklist**:（**请自检，在[ ]内打上x，我们将检视你的完成情况，否则会导致pr无法合入**）

+ - [ ] **设计**：PR对应的方案是否已经经过Maintainer评审，方案检视意见是否均已答复并完成方案修改
+ - [ ] **测试**：PR中的代码是否已有UT/ST测试用例进行充分的覆盖，新增测试用例是否随本PR一并上库或已经上库
+ - [ ] **验证**：PR描述信息中是否已包含对该PR对应的Feature、Refactor、Bugfix的预期目标达成情况的详细验证结果描述
+ - [ ] **接口**：是否涉及对外接口变更，相应变更已得到接口评审组织的通过，API对应的注释信息已经刷新正确
+ - [ ] **文档**：是否涉及官网文档修改，如果涉及请及时提交资料到Doc仓

<!-- **Special notes for your reviewers**: -->
<!-- + - [ ] 是否导致无法前向兼容 -->
<!-- + - [ ] 是否涉及依赖的三方库变更 -->

<!-- bot4-pr-sync-meta
schema_version: 1
source: gitcode
gitcode_repo: openJiuwen/agent-core-java
gitcode_mr: 265
github_repo: openJiuwen-ai/agent-core-java
github_pr: 187
github_branch: sync/pr-265
-->

See merge request: openJiuwen/agent-core-java!265

+++ src/main/java/com/openjiuwen/core/sysop/result/ExecuteCmdBackgroundData.java
@@ -10,21 +10,21 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 /**
- * Data structure for background shell command execution.
- * 
- * @since 0.1.7
+ * Mirrors Python's {@code ExecuteCmdBackgroundData} in
+ * {@code openjiuwen/core/sys_operation/result/shell_operation_result.py}.
  */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class ExecuteCmdBackgroundData {
+
     private String command;
 
     @Builder.Default
     private String cwd = ".";
 
-    private Long pid;
+    private Integer pid;
 
     private String shellType;
 }
